Rename `__data` to `data`
